Fingerprint-Blood-Group-Detection/requirements.txt
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.21.0
opencv-python-headless>=4.5.0
scikit-image>=0.19.0
scikit-learn>=1.0.0
matplotlib>=3.5.0
tqdm>=4.62.0
Pillow>=8.0.0
Flask>=2.0.0
Werkzeug>=2.0.0